Customer.v4 from Integer to Long done

+++ src/main/java/se/backend1/pensionat/entity/Booking.java
@@ -18,7 +18,7 @@ public class Booking {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
 
     @NotNull(message = "Incheckningsdatum krävs")
     private LocalDate checkIn;
